Separate closes #78
