+++ client/System/Desktop.js
@@ -1,47 +1,20 @@
-import { Folder, File, Link, Root } from './FileSystem.js'
+import { Root } from './FileSystem.js'
+import { Reboot } from './System.js'
 import { GameManager } from '../Windows/Window/GameManager.js'
 import { LoginWindow } from '../Windows/Window/LoginWindow.js'
 import { WiFi } from '../Windows/Window/WiFi.js'
-import { Socket } from './Network.js'
 import { Lime } from '../Windows/Window/Lime.js'
 import { Mail } from '../Windows/Window/Mail.js'
 import { InternetBrowser } from '../Windows/Window/InternetBrowser.js'
 import { Store } from '../Windows/Window/Store.js'
 import { ResultWindow } from '../Windows/Window/ResultWindow.js'
-import { JobManager, joblist } from '../Windows/Window/JobManager.js'
-import { MailReciever } from '../Services/Service/MailReciever.js'
-import { JobReciever } from '../Services/Service/JobReciever.js'
+import { JobManager } from '../Windows/Window/JobManager.js'
 import { Explorer } from '../Windows/Window/Explorer.js'
 import { TextInputWindow, YesNoButtonWindow } from '../Functions/InputWindow.js'
 import { AddContextMenu } from '../Functions/contextmenu.js'
 import { Notify } from '../Functions/notify.js'
 
 const UserName = "ANZEN"
-
-let bootmode = "join"
-
-//ココにウィンドウのリストを追加していく
-// const windowList = {
-//     "GameManager": GameManager,
-//     "LoginWindow": LoginWindow,
-//     "WiFi": WiFi,
-//     "Crusher": Crusher,
-//     "Mail": Mail,
-//     "InternetBrowser": InternetBrowser,
-//     "Store": Store,
-//     "ResultWindow": ResultWindow,
-//     "JobManager": JobManager,
-//     "VirusScanner": VirusScanner,
-//     "Excol": Excol,
-//     "Explorer": Explorer,
-//     "Installer": Installer
-// }
-
-//ココに最初から動かすバックグラウンドサービスのリストを追加していく
-const systemServiceList = [
-    MailReciever,
-    JobReciever,
-]
 
 export let WindowManager = {
     /** @type {HTMLElement} */
@@ -53,147 +26,6 @@ export let WindowManager = {
 }
 
 WindowManager.windowarea = document.getElementById("desktop_windows")
-
-/** ウィンドウを呼ぶ
- * @param {String|Object} classfunc
- * @param {String|Number=} window_id
- */
-// export function CallWindow(classfunc, window_id) {
-//     if (!window_id && typeof classfunc == 'string') {
-//         window_id = classfunc
-//     }
-//     if (typeof classfunc == 'string') {
-//         if (windowList[classfunc]) {
-//             CreateWindow(windowList[classfunc], window_id)
-//         }
-//     } else if (typeof classfunc == 'function') {
-//         CreateWindow(classfunc, window_id)
-//     }
-// }
-
-/** サービスを呼ぶ(1個だけ)
- * @param {String|Object} classfunc
- * サービスは1つだけなので識別用IDは無い
- */
-// export function CallService(classfunc) {
-//     if (typeof classfunc == 'string') {
-//         // if (!WindowManager.services[classfunc] && systemServiceList[classfunc]) {
-//         //     WindowManager.services[classfunc] = new systemServiceList[classfunc]();//サービスを作成
-//         // }
-//         new systemServiceList[classfunc]()
-//     } else if (typeof classfunc == 'function') {
-//         // if (!WindowManager.services[classfunc.name] && systemServiceList[classfunc]) {
-//         //     WindowManager.services[classfunc.name] = new classfunc()
-//         // }
-//         new classfunc()
-//     }
-// }
-
-export let Task = {
-    CompletedTask: [],
-    SucceedTask: [],
-    FailedTask: [],
-    Complete: Task_Complete,
-    IsCompleted: Task_IsCompleted,
-    EmitResult: Task_EmitResult
-}
-
-export let Result = {
-    Revenue: 0,
-    SecurityScore: 0,
-    Flag: [],
-}
-
-export let SystemConfigs = {
-    "installed_software": [],      // アプリストアから入れたソフト
-    room: {
-        roomid: "",
-        username: "名無しの社員さん",
-        users: [],
-        myID: "",        // 自分のソケットID
-        status: false,
-        startat: 0,      // 開始時刻
-        duration: 0,     // ゲームの長さ(ms)
-    },
-    connected_wifi: [],
-    Task: Task,
-    Result: Result,
-
-    Packages: {
-        List: [],
-        Install: package_Install,
-        Uninstall: package_Uninstall
-    },
-
-    isWizardClosed: false
-}
-
-function package_Install(name, iconurl, func) {
-    SystemConfigs.Packages.List.push(name)
-    /**@type {Folder} *///@ts-ignore
-    let desktopfolder = Root.GetPath(`/Users/${UserName}/Desktop/`)
-    /**@type {Folder} *///@ts-ignore
-    let progfolder = Root.GetPath(`/Programs/`)
-
-    let tmp = new File(progfolder, name)
-    tmp.icon = iconurl
-    tmp.content = func
-    new Link(desktopfolder, name, tmp).icon = iconurl
-
-}
-function package_Uninstall(name) {
-    SystemConfigs.Packages.List.filter((item) => {
-        return item !== name
-    })
-    /**@type {Folder} *///@ts-ignore
-    let desktopfolder = Root.GetPath(`/Users/${UserName}/Desktop/`)
-    /**@type {Folder} *///@ts-ignore
-    let progfolder = Root.GetPath(`/Programs/`)
-
-    desktopfolder.children.delete(name)
-    progfolder.children.delete(name)
-}
-
-function CreateWindow(func, window_id) {
-    if (WindowManager.windows[window_id]) {
-        WindowManager.windows[window_id].BringToTop();
-    } else {
-        // WindowManager.windows[window_id] = new func();//funcクラスのウィンドウを作成
-        // WindowManager.windows[window_id].window_id = window_id;
-        new func()
-        RefreshTaskbarIcons();
-    }
-}
-
-function Task_Complete(task, failed = false) {
-    if (Task.IsCompleted(task) && !failed) return;//すでにタスクをクリアしていて、失敗でなかったなら
-    else if (!Task.IsCompleted(task)) {//未クリアなら
-        Task.CompletedTask.push(task)
-        if (failed) {
-            if (Task.IsCompleted(task))
-                Task.FailedTask.push(task)
-        } else {
-            Task.SucceedTask.push(task)
-        }
-    } else if (failed) {//クリア済みで失敗したのなら
-        Task.SucceedTask = Task.SucceedTask.filter((item) => {
-            return item !== task;
-        });
-        Task.FailedTask = Task.FailedTask.filter((item) => {
-            return item !== task;
-        });
-        Task.FailedTask.push(task)
-    }
-}
-
-function Task_IsCompleted(task) {
-    let res = Task.CompletedTask.findIndex(t => t === task)
-    return (res === -1) ? false : true
-}
-
-function Task_EmitResult(data) {
-    Socket.emit("taskresult", JSON.stringify(data))
-}
 
 let DesktopIconList_create = [
     { Name: "タスク管理", Iconurl: "/images/jobManagericon.png", Clickfunc: () => { new JobManager() } },
@@ -273,8 +105,8 @@ let MenuIconList_R_join = [
 export let MenuIconList_L;
 export let MenuIconList_R;
 
-function initParameters(){
-    switch(bootmode){
+export function InitParameters(mode){
+    switch(mode){
         case "join":
             DesktopIconList = DesktopIconList_join;
             MenuIconList_L = MenuIconList_L_join;
@@ -293,7 +125,7 @@ function initParameters(){
 
 export function RefreshDesktop() {
     const desktop_icons = document.getElementById("desktop_icons")
-    /** @type {Folder} *///@ts-ignore
+    /** @type {import("./FileSystem").Folder} *///@ts-ignore
     let desktopfolder = Root.GetPath(`/Users/${UserName}/Desktop/`)
     desktop_icons.innerHTML = "";
     let elm_drag
@@ -546,98 +378,35 @@ function createElementFromHTML(html) {
     return template.content.firstElementChild;
 }
 
-export function Boot(_bootmode) {
-    bootmode = _bootmode;
-    initParameters();
-    initFileSystem()
-    RefreshDesktop()
-    RefreshTaskbar()
-}
+// function lsall(dir) {
+//     /** @type {Folder} *///@ts-ignore
+//     if (typeof dir == "string") dir = Root.GetPath(dir).children
+//     else if (!dir) dir = Root.children
+//     dir.forEach(element => {
+//         console.log(element.Pwd())
+//         if (element.isdir) {
+//             lsall(element.children)
+//         }
+//     });
+// }
+// window["lsall"] = lsall
+// window["getpath"] = (path) => { return Root.GetPath(path) }
 
-export function Reboot() {
-    for (let windowid in WindowManager.windows) {//すべてのウィンドウを削除
-        WindowManager.windows[windowid].destroy(true);
+(()=>{
+    const backgrounds = [
+        "/images/backgrounds/hill.svg",
+        "/images/backgrounds/blue.svg",
+        "/images/backgrounds/orange.svg",
+    ]
+    let index = 0
+    setInterval(() => {//壁紙変更
+        index++
+        if (index >= backgrounds.length) index = 0
+
+        setbackground(backgrounds[index])
+    }, 1000 * 60);
+    function setbackground(str) {
+        const desktop = document.getElementById("desktop")
+        desktop.style.backgroundImage = "url(" + str + ")"
     }
-    const bootwindow = document.getElementById("bootwindow")
-    bootwindow.classList.remove("hidden")
-    bootwindow.firstElementChild.innerHTML = "シャットダウンしています…"
-    setTimeout(() => {
-        bootwindow.style.background = "black"
-        bootwindow.firstElementChild.innerHTML = ""
-        initFileSystem()
-        setTimeout(() => {
-            bootwindow.style.background = ""
-            bootwindow.firstElementChild.innerHTML = "ようこそ"
-            setTimeout(() => {
-                bootwindow.classList.add("hidden")
-            }, 2000);
-        }, 1000);
-    }, 2000);
-}
-
-export function InitGame() {//リザルトとかを初期化
-    SystemConfigs.connected_wifi = []
-    SystemConfigs.installed_software = []
-    Result.Revenue = 0
-    Result.SecurityScore = 0
-    Result.Flag = []
-    Task.CompletedTask = []
-    Task.FailedTask = []
-    Task.SucceedTask = []
-    joblist.splice(0, joblist.length)
-
-    if (!SystemConfigs.room.status) {
-        for (let windowid in WindowManager.windows) {//すべてのウィンドウを削除
-            if (windowid != "ゲームマネージャー")
-                WindowManager.windows[windowid].destroy(true);
-        }
-    }
-
-
-    for (let serviceid in WindowManager.services) {//すべてのサービスを停止
-        WindowManager.services[serviceid].destuctor();
-    }
-    systemServiceList.forEach(s => {
-        new s()
-    })
-}
-
-export function StopGame() {//サービス等停止用
-    for (let serviceid in WindowManager.services) {//すべてのサービスを停止
-        WindowManager.services[serviceid].destuctor();
-    }
-}
-
-function initFileSystem() {
-    Root.children.clear()
-    /**@type {Folder} *///@ts-ignore
-    let desktop = Root.NewItem(`/Users/${UserName}/Desktop/`)
-    let document = Root.NewItem(`/Users/${UserName}/Document/`)
-    let picture = Root.NewItem(`/Users/${UserName}/Picture/`)
-    let music = Root.NewItem(`/Users/${UserName}/Music/`)
-    let download = Root.NewItem(`/Users/${UserName}/Download/`)
-    let progfolder = Root.Mkdir("Programs")
-
-
-    DesktopIconList.forEach(icon => {
-        let tmp = new File(progfolder, icon.Name)
-        tmp.icon = icon.Iconurl
-        tmp.content = icon.Clickfunc
-        new Link(desktop, icon.Name, tmp).icon = icon.Iconurl
-    })
-    RefreshDesktop()
-}
-
-function lsall(dir) {
-    /** @type {Folder} *///@ts-ignore
-    if (typeof dir == "string") dir = Root.GetPath(dir).children
-    else if (!dir) dir = Root.children
-    dir.forEach(element => {
-        console.log(element.Pwd())
-        if (element.isdir) {
-            lsall(element.children)
-        }
-    });
-}
-window["lsall"] = lsall
-window["getpath"] = (path) => { return Root.GetPath(path) }
+})()
