+++ client/Windows/Window/Excol.js
@@ -1,6 +1,6 @@
 'use strict'
 import { Window } from "../Window.js"
-import { SystemConfigs } from "../../System/Desktop.js"
+import { SystemConfigs } from "../../System/System.js"
 import { WormVirus } from "../../Viruses/VirusEvents.js"
 import { OpenWorm } from "./OpenWorm.js"
