+++ client/Windows/Window/InitWizard.js
@@ -1,6 +1,6 @@
 'use strict'
 import { Window } from "../Window.js"
-import { SystemConfigs } from "../../System/Desktop.js"
+import { SystemConfigs } from "../../System/System.js"
 import { WizardOverlay } from "../../System/WizardOverlay.js";
 
 let initwizHtml = `
